refactored bottom accessory test for iOS
